requirements.txt
google-genai          # Gemini Embedding 2 for multimodal embeddings
networkx              # hypergraph structure
numpy                 # vector math
faiss-cpu             # similarity search
python-dotenv         # loads .env API key
pillow                # image processing
openai                # OpenAI-compatible SDK to call Nvidia/Qwen via OpenRouter
pymupdf               # fitz — PDF text and image extraction
streamlit             # web UI
pyvis                 # interactive knowledge graph visualization
langsmith             # tracing and observability
datasets              # benchmark datasets from Hugging Face
pandas                # result aggregation and tabular reporting
matplotlib            # benchmark and trace plots
